pyproject.toml
[project]
name = "murkelhausen-app3"
version = "1.8.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "babel>=2.17.0",
    "beautifulsoup4>=4.14.0",
    "cachetools>=6.2.1",
    "django>=5.2.6,<6.0",
    "django-apscheduler>=0.7.0",
    "django-htmx>=1.25.0,<2.0",
    "gcsa>=2.6.0",
    "google-api-python-client>=2.185.0",
    "icalendar>=6.3.1",
    "psycopg>=3.2.11",
    "pydantic>=2.11.9,<3.0",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2025.2",
    "requests>=2.32.5,<3.0",
    "garminconnect>=0.2.25",
    "garth>=0.4.46",
    "sqlalchemy>=2.0",
    "apscheduler>=3.10,<4.0",
]

[dependency-groups]
dev = [
    "ruff>=0.13.1",
    "ty>=0.0.1a21",
]

[tool.ruff]
# Django-specific configuration
target-version = "py313"
line-length = 88
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "migrations",
    "venv",
    "env",
    ".env",
    "build",
    "dist",
    ".github",
]

[tool.ruff.lint]
# Enable Django-friendly rules
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "YTT", # flake8-2020
    "S",   # flake8-bandit
    "BLE", # flake8-blind-except
    "B",   # flake8-bugbear
    "A",   # flake8-builtins
    "C4",  # flake8-comprehensions
    "DJ",  # flake8-django
    "DTZ", # flake8-datetimez
    "T10", # flake8-debugger
    "EM",  # flake8-errmsg
    "ISC", # flake8-implicit-str-concat
    "ICN", # flake8-import-conventions
    "PIE", # flake8-pie
    "T20", # flake8-print
    "PT",  # flake8-pytest-style
    "Q",   # flake8-quotes
    "RSE", # flake8-raise
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
    "TCH", # flake8-type-checking
    "ARG", # flake8-unused-arguments
    "PTH", # flake8-use-pathlib
    "ERA", # eradicate
    "PL",  # Pylint
    "TRY", # tryceratops
    "RUF", # Ruff-specific rules
]

# Ignore specific rules that are common in Django projects
ignore = [
    "DJ001",  # null=True on string-based fields
    "S101",   # Use of assert detected
    "S311",   # Standard pseudo-random generators are not suitable for security/cryptographic purposes
    "B008",   # Do not perform function calls in argument defaults
    "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
    "ISC001", # Conflicts with formatter
]

# Allow specific violations in specific files
[tool.ruff.lint.per-file-ignores]
"settings.py" = ["E501", "F401", "F403", "ERA001"]
"**/settings/*.py" = ["E501", "F401", "F403", "ERA001"]
"manage.py" = ["T201", "PLC0415", "TRY003", "EM101"]
"**/migrations/*.py" = ["E501", "RUF012"]
"family_intranet/repositories/handballnordrhein.py" = ["DTZ007", "PLR2004"]
"family_intranet/repositories/fussballde.py" = ["DTZ007"]
"family_intranet/repositories/gymbroich.py" = ["DTZ005", "DTZ007", "N815", "E501", "S113", "SIM210", "PLW2901", "PLR2004"]
"family_intranet/repositories/mheg.py" = ["DTZ005", "DTZ011", "N815", "E501", "PLR2004"]
"family_intranet/repositories/pihole.py" = ["TRY003", "EM101", "EM102", "DTZ003", "T201", "ERA001", "E501"]
"family_intranet/repositories/google_calendar.py" = ["SIM102", "RET504", "TRY003", "EM102", "DTZ011", "B905", "PLW2901", "PLW0127", "C416"]
"family_intranet/repositories/outlook_calendar.py" = ["PLR0912", "PLR0915", "TRY003", "EM102", "TRY401"]
"family_intranet/repositories/owm.py" = ["TRY003", "EM102", "BLE001", "PLR2004", "E501", "T201"]
"family_intranet/repositories/owm_functions.py" = ["PLR0911", "PLR2004"]
"family_intranet/repositories/owm_models.py" = ["DTZ006", "DTZ012", "E501"]
"core/views.py" = ["DTZ007", "PLR0911"]
"family_intranet/jobs/garmin/*.py" = ["DTZ005", "DTZ007", "DTZ011", "N815", "E501"]
".github/**/*.py" = ["ALL"]

[tool.ruff.lint.isort]
# Django-specific isort configuration
known-first-party = ["core", "family_intranet"]
known-third-party = ["django"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Indent with spaces
indent-style = "space"
# Respect magic trailing commas
skip-magic-trailing-comma = false
# Automatically detect line ending
line-ending = "auto"

[tool.ty]
# ty type checker configuration

[tool.ty.environment]
python-version = "3.13"

[tool.ty.src]
exclude = [
    "*/migrations/*",
    "manage.py",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    ".github",
    "family_intranet/repositories/mheg.py",
    "family_intranet/repositories/pihole.py",
    "family_intranet/repositories/owm.py",
    "family_intranet/repositories/owm_functions.py",
    "family_intranet/repositories/owm_models.py",
    "family_intranet/repositories/google_calendar.py",
    "family_intranet/jobs/garmin/*.py",
]
